requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
gspread==6.0.2
gspread_dataframe==3.3.1
oauth2client==4.1.3
google-auth==2.29.0
